// resources/spa/src/content/generated/contentTypes.generated.ts
export type ContentAccountKey =
  | "merchant_administrator"
  | "merchant_audit"
  | "merchant_branch"
  | "merchant_finance"
  | "merchant_front_office"
  | "merchant_human_resource"
  | "merchant_personnel"
  | "super_administrator";